refactor: split GTFormatter from GT to reduce client bundle size
Extract formatting and locale utilities into a new `generaltranslation/format`
subpath with a lightweight `GTFormatter` class. `GT` now extends `GTFormatter`,
keeping API client methods separate. Client-side packages (`react-core`,
`react`, `next`) import from `generaltranslation/format` instead of the main
entry, avoiding the API client in production bundles.

The dynamic `import('generaltranslation')` for dev-mode runtime translation
lives in `gt-react` only (via `useTranslateMany` hook), keeping `react-core`
bundler-agnostic and safe for React Native / Metro.
