+++ packages/core/tsdown.config.ts
@@ -5,6 +5,7 @@ export default defineConfig(
   createTsdownConfig(
     [
       'src/index.ts',
+      'src/format.ts',
       'src/id.ts',
       'src/internal.ts',
       'src/errors.ts',
