+++ packages/react-core/src/variables/Currency.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { GT } from 'generaltranslation';
+import { GTFormatter } from 'generaltranslation/format';
 import { GTContext } from '../provider/GTContext';
 
 /**
@@ -33,7 +33,7 @@ function Currency({
 }): React.JSX.Element | null {
   const context = useContext(GTContext);
   if (children == null) return null;
-  const gt = context?.gt || new GT();
+  const gt = context?.gt || new GTFormatter();
   let renderedValue: string | number =
     typeof children === 'string' ? parseFloat(children) : children;
   if (typeof renderedValue === 'number') {
