+++ packages/react/rollup.base.config.mjs
@@ -30,11 +30,12 @@ export default {
     postcss(), // Process CSS files
     preserveDirectives(), // Preserve directives in the output (i.e., "use client")
   ],
-  external: [
-    'react',
-    'react-dom',
-    'generaltranslation',
-    '@generaltranslation/supported-locales',
-    '@generaltranslation/react-core',
-  ],
+  external: (id) =>
+    [
+      'react',
+      'react-dom',
+      'generaltranslation',
+      '@generaltranslation/supported-locales',
+      '@generaltranslation/react-core',
+    ].some((ext) => id === ext || id.startsWith(ext + '/')),
 };
